requirements.txt
flask==3.0.0
sqlalchemy==2.0.23
pandas==2.1.4
matplotlib==3.8.2
plotly==5.17.0
jupyter==1.0.0
seaborn==0.13.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
